Remove OneTimePasswordCheckable
Removes the `OneTimePasswordCheckable` as it's now only used in one place
so it's functionality has been inlined into `Reminder`.
Also removes the `otp_category` argument from
`OneTimePassword::Validator` as that argument isn't used.
There's still some weirdness around using `write_attribute` in the
validation, not sure why it was done this way but have left it for the
time being.