Add a form object for reset-claim
When there is a discrepancy with the details returned by DfE Sign In, we
present the option to reset the claim.

Resetting the claim doesn't have any validations but does currently have
a hardcoded next slug. We want to remove this coupling to the next slug
by moving this to a form object.

It is a very simple form object, with its only purpose to allow us to
dynamically define the next slug.