Extract a form for claim submission
This is a bit of a work in progress while we decide what
responsabilities can be moved from CurrentClaim and Claim into this
form.
There's also the question of whether we want to break the mapping
between view and form name. There's a couple of options
1.) Rename this form CheckYourAnswersForm and have it call a claim
  submission job - Not really ideal imo as this form really validates if
  the claim is submittable or not
2.) Don't use the form on check your answers, and if there's an error
  rerender the check your answers page passing the form as
  `current_claim` - A bit weird.
3.) Break the mapping in this case - Least weird imo and I'm not sure
  how useful being able to infer the form from the view in all cases is.

There's quite a few todos in this commit that will be removed in future
commits.