Extracts job for submitting form
Moves the logic for submitting a claim off of the claim and the
eligibility into a ClaimSubmission job / service.

It's a little tricky as we're working with multiple claims, however once
we refactor where a claim is created this can be tidied up.
Also `submittable?` needs to remain as a method on the Claim as it's
currently used by `PageSequence` to determine when to redirect the user
to the check answers screen. Again once we've refactored creating a
claim out of the journey we'll be able to remove this submission related
method off the Claim.