Introduces journey session model
As part of the efforts to simplify multi claim handling we want to move
away from creating a claim at the start of the journey and instead only
create the claim at the final step where the teacher submits their
answers.

This commit introduces a new model `Journeys::Session` which will be
used to store the answers entered by the teacher. The answers will be
stored in a json column on the new model, once validated by the form
objects.

`Journeys::Session` is currently the best name we could come up with but
we're all ears for any better suggestions.