Make journey_session available in forms
We'll soon be writing answers to the journey session as well as to the
form object. This commit makes the journey session available to, and
required in, the forms.
As we use a form object in the slug sequence and the slug sequence is
called by the page sequence we've had to pass the journey session
through those classes as well.