Create a claim from journey session
As part of migrating away from `CurrentClaim` we want to create a
`Claim` from the `Journeys::Session` model, which is where we'll be
storing the answers from the various forms in the wizard.
We're starting this migration by ensuring creating a `Claim` from the
`Journeys::Session` works. As we haven't updated all the form objects to
write to the `Journeys::Session` we have introduced a temporary shim.
This temporary shim class trys to get an answer from the
`Journeys::Session` and falls back to the `CurrentClaim` if there is no
journey session answer. As each form is updated to write to the
`Journeys::Session` we can update a line in
`ClaimJourneySessionShim#answers` until this class does nothing at which
point it can be removed and the `Journeys::Session` can be queried
directly.

The `SubmissionController` now deletes all `Claim` in the `CurrentClaim`
as `CurrentClaim` no longer submits a `Claim`. This is done in the
controller to avoid passing `CurrentClaim` as an argument to the
`ClaimSubmissionForm`, and when we no longer use `CurrentClaim` this
line can be deleted.

The `ClaimSubmissionForm` has attributes for the `Claim` and the
`Eligibilities`, these attributes are set from the `Journeys::Session`
which allows us to use `validates` to check the form is submittable. We
could avoid defining these attributes on the the `ClaimSubmissionForm`
and pull the attributes directly off the `Journeys::Submission` and
instead just use `validates`, I'm unsure which approach is neater.

We've also added a new column to the `Claim`, `journeys_session_id`
which allows us to associate the `Claim` with the `Journeys::Session` it
was created from. This association lets us know that a
`Journeys::Session` has been submitted, to avoid duplicate submissions,
and it will assist in replacing the `Eligibility` record with a plain
Ruby object. Unfortunatley we have to make this `belongs_to` optional as
there are existing claims in the database and `Claim`s are still created
at the start of the journey. Once we've stopped creating `Claim`s at the
start of the journey we may want to consider backfilling the existing
`Claims` on production with `Journeys::Session` records and make this
`belongs_to` association non optional.