Always return `true` when calling `Form#persisted?`
This is another step to avoid coupling a generic form object to a specific AR model.

As a reminder, we keep `#persisted?` so Rails can make an informed decision to use PATCH
instead of POST when submitting a form. By definition, our form objects are and will always be
persisted "somewhere" (most likely not the DB in the future)