Add a form object for selecting an address
When the user is adding their home address, we have a hard-coded address
selection form that can be migrated to a form object.

This change allows us to remove the AddressDetails concern from the
controller as the functionality it adds is moved into the form objects.

Also, the rescue block in the show action of the ClaimsController is no
longer called from anywhere and we can safely remove it too.