Load values from journey session
If the journey session has a value in the answers hash prefer that when
loading the initial attribute values for a form.
Note that `Journeys::Session#answers` is intended to be a "flat" data
structure, not differentiating between claim and eligibility attributes
as it's just concerned with storing answers provided by the user, not
what they're mapped to in the data base.