Use attributes with journey session
Working with an attributes hash is a bit cumbersome as one has to be
careful with updates to avoid overwriting existing answers, eg
`journey_session.update!(answers: { somekey: "some value"})` overwrites
all answers. Additionally working with `[]` to access answers is tricky
as typo'd keys are hard to detect. Additionally some information is
accessed on a claim pre and post submission so we want an object that
can mimic the expected api (eg when working with DqtRecord).
Further we need typecasting of some attributes such as Academic year,
which add a lot of noise when serialising / deserialising to / from a
hash.