Pass eligibility to dqt record
Pass the eligibility into dqt record rather than the claim. Most of the
attributes dqt record references are on the eligibility rather than the
claim. ECP dqt records access academic_year off the claim so we need to
delegate that on ECP eligibiltiy. Passing in the eligibility will make
supporting passing in journey session answers easier. Additionaly we
update the dqt record specs to test against an open struct rather than
the eligibility as we'll want dqt record to work with both eligibility
and journey session answers so we'll test against a generic api.