Migrate signin or continue form to use session
This commit updates the sign in continue form to write it's answers to
the journey session model. The sign in or continue form sets the
`teacher_id_user_info` attributes and the dqt teacher attributes.
Migrating these to the claim has required some substantial changes as
this info is used everywhere.

Of note we've left some attributes in the sign in form still writing to
the current claim, specifically the personal details attributes to avoid
having to update that form at the same time as doing this form.

We should consider backing the answers with a table rather than a json
column, there's quite a few places where we want to call `save` on the
answers directly rather than have to call save on the journey session.