Migrate personal detail form to write to answers
Updates the personal details form to write to the session answers.
This has let us clean up some of the current claim handling in the
related sign in or continue form.
We've moved handling reseting depenent answers into the personal details
form as this is where the personal details are changed. We could
consider moving this on to the session model in a similar manner to how
is done on the claim but for now I think the form is a better place for
this.