Add reminder specific email verificaiton form
Previously reminders email verification form inherited from the main
email verification. This commit copies the email verification form code
to the reminder email verification form as we want to change email
verification form to write to the journey session.
This is duplicating some code but duplication is better than the wrong
abstraction.
The specs for the reminder email verification form stub `update!` and
just assert it is called with the expected arguments, this is something
we should fix in future and instead test the records are updated as
expected.