Fix test
We want to avoid just testing that `update!` has been called and instead
test that the model has actually been updated correctly.
Just testing that `update!` has been called fails to catch bugs if the
model we're saving is changed, or if we change the implementation of
`update!` in the form base class.