Update mobile forms to write to journey session
Updates the various mobile forms to write to the journey session and
updates the call sites to read this information from the journey
session.
We've had to change the order of things in the claim submission form
slightly as the `mobile_number_verified` validation not longer creates a
claim. When `claim_is_eligible` validation runs we need to have the
claim set up and associated with the `main_eligibility`, as this is no
longer done by the `mobile_number_verified` validation we set the claim
up in the constructor.