Migrate bank account form to write to session
Updates the bank account forms to write their answers to the session.
We've also added a new trait to the answers factories "submittable". As
we migrate more code to writing to the answers we can just update this
trait rather than updating the individual specs.
We've moved normalising the users answers into the form object rather
than on the model as data normalisation seems like a concern of the
form.