Fix changes logic
When we load session attributes from the database and populate the
answers object's attributes that counts as a change. To avoid this false
positive we clear the changes information after initialize.
While we're only using this changes logic during the migration from
current claim to journey session, this issue highlights the trickiness
of working with a hash backed column rather than a table backed column.
Once the migration work is completed, and everything is writing to the
session object, we should consider the maintainability of this approach
and see whether a table backed approach would meet our needs better.