Move claim school to answers
Moves the claim school to the session answers.
Claim school is used in a bunch of eligibility checks, so we now need to
be able to check eligibility of the answers and of the eligibility, and
as we're part way through the migration away from current claim we need
to be able to check the eligibility of the combination of both! To do so
we've introduced a new class
`Policies::StudentLoans::EligibilityChecker` which accepts either a
session answers or a claim and can be queried for eligibility status. We
can also pass the "shim" to this class for when we need to check the
eligibility across a combination of answers and current claim.
Eventually this class should probably subsume
`lib/ineligibility_reason_checker` too but that's something for a future
commit.