qualificaton details to use write to answwers
As we're still pulling the `qualification`, `eligible_itt_subject` and
`itt_academic_year` form the claim, we still need to write
`qualification_details` to the claim so that the check in reset depenent
answers is hit which removes deleting the depenencies.
Hopefully we'll soon be able to remove this reset dependent answers
stuff and just nullify the answers directly in the forms.