Store the nqt_in_academic_year_after_itt on the journey session
We want to store the value of the nqt_in_academic_year_after_itt on the
journey session.

There are places in the eligibility check that rely on this value being
stored on the eligibility record. Removing this expectation requires
changing how the eligibility is calculated.

That seems to be beyond the scope of this particular change and will
best be addressed directly rather than as a side-effect of this PR.