Update the itt-year-form to write to answers.
itt_academic_year is used through out the code base to calculate
eligibility, to facilitate this migration we've had to make some
changes.

Expose eligible_now_and_sorted, eligible_now, eligible_now_or_later on
the ClaimSubmissionForm
These methods were used on the EligibilityConfirmedForm and as we now
can't call these methods on the current claim as it doesn't have
itt_academic_year set we call them on the ClaimSubmissionForm which
already handles checking the answers and claim for eligibility

Update the EarlyCareerPayments and LevelingUpPremiumPayments
PolicyEligibilityChecker class to take answers rather than the shim in
their initialiser. These classes only call method on the answers rather
than the shim and are easier to call indecently of the journey
eligibility checkers if they work with just answers. (Also the
JourneySession is just an artefact of how we're persisting the data, all
we really care about is the answers, so we want to work directly with
them as much as possible in case we decide to change to a table backed
answers model).

Update lib/ineligibility_reason_checker to work with answers rather than
a current claim. As the details required to calculate eligibility will
be found on the answers model, this class needs to work with answers
(while we're migrating we actually pass this class the shim).

Update lib/journey_subject_eligibility_checker.rb to work with answers.
We've also moved the subject symbols method out out the app helpers and
into a class method on journey_subject_eligibility_checker.rb. The main
change on this class is how we find the potentially still eligible
policies. Previously we did this be querying the various claims on the
current claim, however as we can no longer do this we loop through the
policies in the journey. As this class is only used for the additional
payments journey and is closely tied to the domain model we should move
it out of the lib directory.

Update the slug sequence to check various eligible conditions against
the PolicyEligibilityChecker classes. There's still some checks agains
the eligibility but these can now be switched over to check the
appropriate PolicyEligibilityChecker as we progress with the rest of the
migration