Set student loans on session
There's a call back that sets student loans data on the claim, we need
to change this to set the student loans data on the session. The
ClaimStudentLoanDetailsUpdater is also used for setting data on
submitted claims so we've introduced an answers specific version and a
presenter class to share the data between the claim and answers
versions.

We've also removed setting the `submitted_using_slc_data` on claims from
the `ClaimStudentLoansDetailsUpdater` as this class is now only called
on submitted claims.