Migrate employed directly to the journey session
We want to store the employed_directly answer on the journey session and
move it away from the eligibility object.

This does impact how the ineligibility reason is calculated. For now I
opted to include the answers to the ineligibility checker class. This
seemed like the simplest change for the short-term.

Other work is progressing to move the logic from
`IneligibilityReasonChecker` to the journey session.