Remove loading answers from claim
Removes the hacks we had to put into the journey session to check if it
contained the answer to a question.
We've left setting attributes from the claim on the reminders forms as
they still write some otp data to the claim / do some weired stuff with
writing data from the session to the form.