Remove claim as a form argument
Now the forms no longer reference the current claim we can remove it as
an argument.
Of note we've had to hack the reminders form a bit. The reminders
controller passes in the reminder to the form as the claim argument.
We've removed the "current_data_object" method as it's too much
abstraction and added some methods to the reminder forms to still let
them work with the claim, which is really a reminder. The reminders
forms probably shouldn't share a base class with the usual forms but
unpicking that is a future exercise. These hacks will hopefully be
addressed in https://dfedigital.atlassian.net.mcas.ms/browse/CAPT-1719