Remove claim_in_progress? view helper
This helper was defined slightly differently from the same method in the
controller, making switching out the definition for a claim in progress
tricky. This adds a new method to the controller
`eligible_claim_in_progress?` and exposes `claim_in_progress?` as a
helper from the controller to the view.
We've added a separate line for helper definition to make a rebase
easier, this will be moved inline with the other helpers in a future
commit!