Only support a single claim at a time
If a teacher starts a claim then switches journey the previous behaviour
was to show the teacher the option to continue the claim they have in
progress or start a new claim and lose their progress. Previously the id
for both claims was stored in the session under a single key. The
journey sessions were stored under a sepearte key for each journey.
As we want to match the existing behaviour we now need to know about
what other journey sessions are in progress.
This has introduced a bit of complexity and ideally we should support
multiple in progress claims but that is something we need to discuss
with the product team. (My vote is to support multiple in progress
claims as that requires less code!)