Initial IRP journey wire up
This commit adds the initial set up for IRP along with a some feature
test.
So far the IRP journey lets a user click "Start now" and then
shows the check your answers page where they can submit a claim.
The feature test has some temporary hacks, where we update the database
model, this is just required to get the answers in a state that passes
the validations on the claim. We'll be removing the
`teacher_reference_number` question and (probably) the `payroll_gender`
question from the IRP journey but they're currently required.