Show the correct name for the previous journey
The `current_journey_routing` name shows the name of the journey we're
switching too. We need to take a look at the journey the user came from
which will be the `first` journey session.

Due to how we clear the session when the user starts a new journey there
will only ever be two journey sessions in the `journey_sessions`
collection. In this specific case on the existing session page, there is
only one journey session in the journey_sessions collection (as the
session for the journey we're considering switching too hasn't been
created)