Add start date form
Adds the form to capture the start date. Of note is the use of
`_assign_attributes`. ActiveModel doesn't support multi part parameters
so we need to include `ActiveRecord::AttributeAssignment` to get them.
This behaviour should probably be upstreamed into the base Form class
and the personal details form migrated to set it's mutli part parameters
with this approach, though changing the attribute assignment in the base
form to support this felt out of scope of this commit.

Also this commit doesn't add the eligibility checking for invalid
contract dates. We're waiting on the product team to get back to us with
what this should be. When they have we'll enable the feature tests for
the ineligibility.