Remove trainee pages
IRP is no longer going to be eligible for trainee teachers. This commit
removes the trainee teacher screens we'd migrated so far and updates the
eligibility to make trainee teachers ineligible.
We've also removed the resetting of depenent answers from the
application for as now if the user changes their answer to any of the
options they'll be unable to proceed with the reset of the journey.