Adds the entry date form
As in claim the entry date form writes to the date_of_entry attribute.
We also reset the date of entry if the start date changes as the
eligibility calculation for date of entry depends on the start date.

Of note there doesn't seem to be a validation in claim that the date of
entry isn't _after_ the start date of the contract. I'm not sure if
that's a business rule we want to enforce or not. There is a validation
that the date of entry isn't in the future.