Setup the admin to display IRP claims
In order to be able to display the IRP claims in the admin, there are
some basic methods that need to be available.

This change ensures the minimum implementation for displaying claims
without being concerned about the correctness of the required methods.
Specifically the `award_amount` is a placeholder.

In order to ensure we can successfully check for duplicate claims, I
made an assumption that we would use passport number rather than teacher
reference number as a unique identifier.

I don't think claimants for IRP will have TRNs, therefore we need some
other way of matching.

We can change this assumption easily in the future but for now this
allows us to ship the admin changes and keep moving forward.