Add nationality screen
We want to display the nationality in the personal details section of
the journey. However claim's don't have a column for nationality, and as
we're using the claim for shared  details between journeys, we've
decided to write the nationality to the eligibility column, even though
it's not part of the eligibility criteria.

We've added a new initializer to store the available nationality
options, even though the list is only used in the form it's a bit
awkward keeping it in there, and an initializer was the approach taken
in the original IRP app.