Add the passport number screen
Adds a screen for the passport number. Again like, the nationality, this
is a question we show in the personal details part of the journey but we
store it on the eligiblity, as this is the only journey that asks this
question and the claim is shared with all journeys.