Add employment details form
Adds the employment details page as per IRP. We've named the fields the
same as in IRP. We could probably provide a school search box for this
but that's something for a future iteration.
Have also added the school details to the PII section to be on the safe
side as they might be enough to identify an claimant in combination with
some of the other non pii answers.