Fix visiting landing page for same journey
Previously if a claimant started a journey then visited the landing page
for that journey and choose to continue with their existing session we'd
throw an error. The `other_journey_sessions` array contains the journey
sessions for the other journeys, which if you're still on the same
journey would be empty, so we need to check the current journey session
(the session for the journey we're currently on).

How we handle managing which journey you're on is subtle and spread
across many places in the controllers / concerns, the overall approach
likely needs revisiting.