Extract claim scrubber
Extracts the scrubbing logic from the claim_personal_data_scrubber into
it's own class. Claim personal data scrubber was responsible for both
finding the claims to scrub and scrubbing them. This commit moves the
responsibility for scrubbing the data to a separate class as we're going
to want to scrub different attributes depending on the policy we're
scrubbing claims for.
The claim data scrubber preformed and `update_all` when nullifying the
attributes on the claim, we've preserved this behaviour of skipping
callbacks / validations by using `update_column`.
Previously we scrubbed all the claims inside a single transaction, this
commit changes the behaviour slightly in that we open a transaction for
each claim we scrub. This should slightly improve throughput (and makes
sense based on how the code is called) but slightly changes the
behaviour in that if there is an error when removing personal data from
a claim previous claim's changes won't be rolled back. Likely not an
issue but worth calling out.