Retain NI and name on IRP claims for 2 years
Updates the IRP claim scrubber to retain the national insurance number
and name details for 2 years before scrubbing them too.

When this method is called by the job we first call super to remove the
not retained PII fields from the claim then we deal with scrubbing any
retained PII claims.