Don't use TRN in match
When validating that all claims for a given payment belong to the same
claimant we can no longer use TRN as some journeys don't capture a trn.
Eg if we have two claims one for ECP and one for IRP then the ECP claim
will have a TRN but the IRP claim will not and so this check will fail.
After discussion with the policy team we've decided to use the NINO as
part of this check. There is some concern that applicants may mistype
their NINO, if this proves an issue in practice we can just drop the
NINO from the list of compared attributes and rely on the matching bank
details.

We've also removed the check on top ups that a candidate matching
claim's policy is in the `OTHER_CLAIMABLE_POLICIES` list of the claim
we're checking against. This check was in place to support some polices
having trns and some not having trns and as we're now matching on NINO
it is no longer needed.