Add placeholder rejection reason
When running the personal data scrubber spec we create a `Decision`
instance. A `Decision` requires a rejection reason to be valid. The
rejection reason is taken from the `Decision`'s `Claim`'s policy'. To
avoid needing to remove / disable the test, and risk not reintroducing
it, we've added a temporary `placeholder` rejection reason (the other
option was stubbing the `ADMIN_DECISION_REJECTED_REASONS` constant in
the spec). When we get round to adding the rejection reasons for FE we
can remove this place holder and the tests will work with whatever is
the first value in the `ADMIN_DECISION_REJECTED_REASONS` constant.