Pass parameters to sesson creation
On some journeys we'll want to be able to set answers based on url
parameters when starting a journey. This commit adds support for setting
answers when creating a session introducing a journey session form. The
session form is passed the parameters and subclasses of the form permit
vaiorus parameters based on their attributes.