Add initial provider journey happy path
This commit adds the happy path for the FE provider journey.
This journey starts with the provider visiting a link (that will be
emailed to them) to begin approving a claim. The provider is redirected
to DfE sign in, and then if authroised shows the claim to approve.

This commit is just the happy path, so we haven't added any of the logic
around checking DfE sign in permissions. We currently allow any DfE sign
in user to approve any claim. We've also not added the questions to the
provider needs to answers to the claim form and we don't change any
state when the form is submitted. This is very MVP, provider just checks
the decleration box and hits submit!

We've also added a landing page which we can link to from the manage
service page in DfE sign in. Currenlty the journey requires a `claim_id`
to work so following the link from DfE sign in just shows the user a
message that they can't access the service like that.

While the approach of representing approving claims as a "journey"
works, it requires us to setup a few models the claims journey expects
to be present that our specific use case doesn't require. There's also a
requirement, on the distant horizon, to have an index page of claims
that require approving by a given approver which might be tricky to fit
into the journey based approach. We may want to consider moving to [a
different approach](https://github.com/DFE-Digital/claim-additional-payments-for-teaching/pull/3074)