Add school to FE eligibility
As part of the FE provider journey we need to check the ukprn on the
claim's school. Claims get their school from their eligibility's
`current_school` method, so we need to make sure that is set for FE
eligibility.
We've named the field `school_id` rather than `current_school_id` as the
FE journey stores the school in the answers `school_id` column. We alias
`school` to `current_school` so the `Claim#school` method still works.