Require a matching school to verify claims
Only users who are signed in to DfE sign in with an organisation that
matches the school the claim is for can verify claims.

This commit introduces the concept of "AuthorisedSlugs" to guard certain
slugs in a journey.
To restrict access to certain slugs a journey needs to define
* `SlugSequence#requires_authorisation?` which accepts a `slug` and
  returns a boolean representing whether or not that slug requires
  authorisation

* `SlugSequence#unauthorised_path` which accepts a `slug` and a
  `failure_reason` and returns a url to send users who fail
  authorisation

* `Authorisation` a class that will be initialized with the journey
  session `SessionAnswers` and the `current_slug` and needs to implement
  the method `Authorisation#failure_reason` returning a string or
  symbol representing why the journey session failed authorisation

As `Authorisation` is only checked if the `current_slug` requires
authorisation journeys that don't restrict access to slugs don't need to
implement and authorisation class. Additionally the `PageSequence`
checks if the `requires_authorisation?` method exists on the
`SlugSequence` before trying to call it, so journeys without restricted
slugs don't need to define anything extra on their `SlugSequence`.

We've added `unauthorised` to the `PageSequence::DEAD_END_SLUGS` list so
we don't generate back links on the unauthorised page and we avoid
getting redirected, in an infinite loop, back to the slug we're
unauthorised for (we don't updated the visited slugs list). A future
improvement would be to move `DEAD_END_SLUGS` into the `SlugSequence`
but I'm trying to avoid changing too much!

On the FE provider journey specific changes:
We've added the OmniauthCallbackForm to handle mapping the omniauth
payload to the journey session's fields. I'd considered splitting out
the actually using the AuthorisedSlugs feature for the FE provider
journey into a separate commit but decided it's better to keep the
context of how it's used in the commit that introduces it.