Allow api user to access response
If a user doesn't have access to the service the response code will be a
404. `https://github.com/DFE-Digital/login.dfe.public-api?tab=readme-ov-file#get-user-access-to-service`
As part of the FE provider work we want to be able to tell if a user has
service access, so we need to be able to check the response status
rather than throwing an error.
Accessing the `role_codes` method without service access will still
throw an error to preserve the existing behaviour.