Check service access
We require the provider to have access to the service in DfE sign in
before they can verify claims. If they don't have access to the service
we want to show them instructions on how to request service access.
We can determine service access by whether or not the request to get
their organisation roles for the services is successful or not,
https://github.com/DFE-Digital/login.dfe.public-api?tab=readme-ov-file#get-user-access-to-service
We've updated the omniauth_callback_form to request their service roles
and store whether the request was successful or not (we'll store the
role information in a future commit).