Restrict admins from verifying claims
A business rule is to not allow claim admins to verify claims.
We need to check for the admin role codes before checking for the
verify claim role code to handle the case where an admin without
the verify claim role tries to verify a claim, otherwise they'd see the
general "you don't have access" unauthorised page.