Add fixed contract assertions to verification form
When displaying the claim verification form to the verifier we need to
ask them yes/no questions about the claimant's eligibility answers.
The questions we ask depend on the eligibility answers. We model the
questions we need to ask as "AssertionForm" instances and, in a future
commit, choose the list of assertions the verifier needs to make based
on the answers given by the claimant.
This commit doesn't address how we're going to store the verifiers
answers, that is a decision for later!
When displaying the `subjects_taught` question we need to show the list
of courses the claimant has indicated that they teach. This list of
courses is stored in a json array on the eligibility model. To make the
list of courses easier to work with, we've introduced a `Course` object
containing the course and subject strings and handling generating the
description of the course, which may include a link.

There's also some stubbed out methods on the eligibility, these will be
removed once the eligibility model piece has been completed and we
rebase this commit.