Store verification details
We need to capture information about who verified the claim and what
checks were shown to present in the admin section off the app.
Usually we'd want to model this by introducing a couple of models
something like:

```mermaid
erDiagram
    Claim ||--|| Verification : "has one"
    Verifier ||--o{ Verification : "has many"
    Verification ||--o{ Assertion : "has many"
```

However, given we're eventually planning on removing the eligibility
models and making them JSON fields on the claim it seemed more in
keeping with the plans for the app to store verification as a JSON
field. Additionally we're unlikely to be querying verification
information and we're unlikely to be updating it, so the downsides of a
blob of JSON shouldn't be too bad. We can always migrate to a relational
approach if needs be.

We've stored the verification on the eligibility rather than on the
claim as, currently, this verification step is policy specific. It may
be needed on other journeys, at which point it shouldn't be too tricky
to move this field to the claim.

Serializing / deserializing this JSON structure into objects is left for
a future commit when we have more detail on what we need (may not
require a serializer.