Handle multiple inflight session
If the provider visits a second magic link while the first is still in
progress make sure they're on the journey for the latest link they
clicked.
It's tricky to change the existing session in progress logic (we need to
write a url param to the journey session) and it doesn't make sense for
this journey, as there's only a single screen so there isn't any
information in progress to lose.
We piggy back off the existing `start_with_magic_link?` logic that skips
the session in progress page, and as we don't set `code` or `email`
params the `handle_magic_link` method is also skipped.