Email provider
When the claimant submits their application we need to send an email out
to the provider to verify the claim.
This commit adds the mechanism to send the verification email.

Currently we've stubbed the verification deadline, to submission date +
1 week, while we wait on finding out what the actual deadline should be.
We've also set the email template for the claimant email to a,
temporary, empty template so the mailer doesn't error.