Update the verification due date
We want the verification due date to be 2 weeks after the claim is
submitted.
This commit also moves the knowledge of the due date to the policy (it's
a policy concern not eligibility concern) and updates the formatting of
the dates in the emails to be consistent with our other emails.