Add provider verification admin task
Adds the admin task for reviewing the provider's verification of the
claimant's claim.
This task is currently a manual check by the admin but in a future
iteration will be replaced with an automated check. There isn't a design
for the manual check yet but all the other checks use a yes / no radio
button so we've gone for that, the copy for the admin question may be
suject to change.
There's a bit of awkwardness in presenting the verification table as
there questions the provider signs off on are slightly different to
those shown in the admin task. Rather than clutter the
admin_tasks_presenter, which will handle showing information about other
tasks, the admin task presenter calls a specific verification presenter
to handle displaying the verification table.
Some of this code feels a bit awkward, maybe we need to consider moving
the verification json into and object.